Add files via upload
Made a program in which the computer randomly chooses a number between 1 to 10, 1 to 100, or any range.
Then give users a hint to guess the number. Every time the user guesses wrong, he gets another clue, and his score gets reduced. 
The clue can be multiples, divisible, greater or smaller, or a combination of all.